UCS-108 : initial UCSRest/Core split
