+++ PEPRest/pom.xml
@@ -14,15 +14,14 @@
 	<version>2.0-SNAPSHOT</version>
 
 	<properties>
-		<jclouds.version>1.9.1</jclouds.version>
-		<springfox.version>2.2.2</springfox.version>
-		<swagger.version>1.5.3</swagger.version>
+		<springfox.version>2.5.0</springfox.version>
+		<swagger.version>1.5.22</swagger.version>
 	</properties>
 
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.1.4.RELEASE</version>
+		<version>2.1.6.RELEASE</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 
@@ -73,6 +72,12 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
+			<exclusions>
+				<exclusion>
+					<groupId>com.vaadin.external.google</groupId>
+					<artifactId>android-json</artifactId>
+				</exclusion>
+			</exclusions>
 		</dependency>
 
 		<dependency>
