+++ PEPRest/src/main/resources/application.properties
@@ -8,8 +8,8 @@ server.port=9999
 pep.id=0
 pep.uri=http://localhost:9999
 pep.revoke-type=HARD
-pep.policy-path=policies/PolicyWatch.xml
-pep.request-path=requests/Request.xml
+pep.policy-path=xmls/PolicyWatch.xml
+pep.request-path=xmls/Request.xml
 pep.api-status-changed=onGoingEvaluation
 
 ucs.uri=http://localhost:9998
