+++ PIPReader/src/test/resources/application.properties
@@ -23,7 +23,7 @@ requestType=<Request xmlns="urn:oasis:names:tc:xacml:3.0:core:schema:wd-17" Comb
 
 subjectPip={\
     "id" : "PIPReaderEnv",\
-    "className" : "it.cnr.iit.ucs.pipreader.PIPReader",\
+    "name" : "it.cnr.iit.ucs.pipreader.PIPReader",\
     "attributes" : [ {\
         "CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
         "EXPECTED_CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
@@ -37,7 +37,7 @@ subjectPip={\
 
 resourcePip={\
     "id" : "PIPReaderEnv",\
-    "className" : "it.cnr.iit.ucs.pipreader.PIPReader",\
+    "name" : "it.cnr.iit.ucs.pipreader.PIPReader",\
     "attributes" : [ {\
         "CATEGORY" : "urn:oasis:names:tc:xacml:3.0:attribute-category:resource",\
         "EXPECTED_CATEGORY" : "urn:oasis:names:tc:xacml:3.0:attribute-category:resource",\
@@ -51,7 +51,7 @@ resourcePip={\
 
 actionPip={\
     "id" : "PIPReaderEnv",\
-    "className" : "it.cnr.iit.ucs.pipreader.PIPReader",\
+    "name" : "it.cnr.iit.ucs.pipreader.PIPReader",\
     "attributes" : [ {\
         "CATEGORY" : "urn:oasis:names:tc:xacml:3.0:attribute-category:action",\
         "EXPECTED_CATEGORY" : "urn:oasis:names:tc:xacml:3.0:attribute-category:action",\
@@ -65,7 +65,7 @@ actionPip={\
 
 environmentPip={\
     "id" : "PIPReaderEnv",\
-    "className" : "it.cnr.iit.ucs.pipreader.PIPReader",\
+    "name" : "it.cnr.iit.ucs.pipreader.PIPReader",\
     "attributes" : [ {\
         "CATEGORY" : "urn:oasis:names:tc:xacml:3.0:attribute-category:environment",\
         "EXPECTED_CATEGORY" : "urn:oasis:names:tc:xacml:3.0:attribute-category:environment",\
@@ -79,7 +79,7 @@ environmentPip={\
 
 missingCategory={\
     "id" : "PIPReaderEnv",\
-    "className" : "it.cnr.iit.ucs.pipreader.PIPReader",\
+    "name" : "it.cnr.iit.ucs.pipreader.PIPReader",\
     "attributes" : [ {\
         "CATEGORY" : "",\
         "EXPECTED_CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
@@ -93,7 +93,7 @@ missingCategory={\
 
 missingAttributeId={\
     "id" : "PIPReaderEnv",\
-    "className" : "it.cnr.iit.ucs.pipreader.PIPReader",\
+    "name" : "it.cnr.iit.ucs.pipreader.PIPReader",\
     "attributes" : [ {\
         "CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
         "EXPECTED_CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
@@ -107,7 +107,7 @@ missingAttributeId={\
 
 missingExpectedCategory={\
     "id" : "PIPReaderEnv",\
-    "className" : "it.cnr.iit.ucs.pipreader.PIPReader",\
+    "name" : "it.cnr.iit.ucs.pipreader.PIPReader",\
     "attributes" : [ {\
         "CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
         "EXPECTED_CATEGORY" : "",\
@@ -121,7 +121,7 @@ missingExpectedCategory={\
 
 missingDataType={\
     "id" : "PIPReaderEnv",\
-    "className" : "it.cnr.iit.ucs.pipreader.PIPReader",\
+    "name" : "it.cnr.iit.ucs.pipreader.PIPReader",\
     "attributes" : [ {\
         "CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
         "EXPECTED_CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
@@ -135,7 +135,7 @@ missingDataType={\
 
 missingFilePath={\
     "id" : "PIPReaderEnv",\
-    "className" : "it.cnr.iit.ucs.pipreader.PIPReader",\
+    "name" : "it.cnr.iit.ucs.pipreader.PIPReader",\
     "attributes" : [ {\
         "CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
         "EXPECTED_CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
@@ -148,7 +148,7 @@ missingFilePath={\
   }
 
 malformedInput="id" : "PIPReaderEnv",\
-    "className" : "it.cnr.iit.ucs.pipreader.PIPReader",\
+    "name" : "it.cnr.iit.ucs.pipreader.PIPReader",\
     "attributes" : [ {\
         "CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
         "EXPECTED_CATEGORY" : "urn:oasis:names:tc:xacml:1.0:subject-category:access-subject",\
