+++ UCSCore/src/main/java/META-INF/additional-spring-configuration-metadata.json
@@ -1,0 +1,107 @@
+{"properties": [
+  {
+    "name": "ucs.pip-list[0].id",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.pip-list[0].id'"
+  },
+  {
+    "name": "ucs.core.uri",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.core.uri'"
+  },
+  {
+    "name": "ucs.core.journal-dir",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.core.journal-dir'"
+  },
+  {
+    "name": "ucs.context-handler.name",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.context-handler.name'"
+  },
+  {
+    "name": "ucs.request-manager.name",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.request-manager.name'"
+  },
+  {
+    "name": "ucs.session-manager.db-uri",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.session-manager.db-uri'"
+  },
+  {
+    "name": "ucs.policy-decision-point.journal-dir",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.policy-decision-point.journal-dir'"
+  },
+  {
+    "name": "ucs.obligation-manager.name",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.obligation-manager.name'"
+  },
+  {
+    "name": "ucs.pep-list[0].name",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.pep-list[0].name'"
+  },
+  {
+    "name": "ucs.policy-administration-point.path",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.policy-administration-point.path'"
+  },
+  {
+    "name": "ucs.request-manager.api-remote-response",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.request-manager.api-remote-response'"
+  },
+  {
+    "name": "ucs.context-handler.uri",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.context-handler.uri'"
+  },
+  {
+    "name": "ucs.request-manager.active",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.request-manager.active'"
+  },
+  {
+    "name": "ucs.session-manager.name",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.session-manager.name'"
+  },
+  {
+    "name": "ucs.policy-decision-point.name",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.policy-decision-point.name'"
+  },
+  {
+    "name": "ucs.policy-administration-point.name",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.policy-administration-point.name'"
+  },
+  {
+    "name": "ucs.pip-list[2].attributes[0].CATEGORY",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.pip-list[2].attributes[0].CATEGORY'"
+  },
+  {
+    "name": "ucs.pip-list[1].attributes[0].CATEGORY",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.pip-list[1].attributes[0].CATEGORY'"
+  },
+  {
+    "name": "ucs.pip-list[0].attributes[0].DATA_TYPE",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.pip-list[0].attributes[0].DATA_TYPE'"
+  },
+  {
+    "name": "ucs.pip-list[1].additional-properties.KEY",
+    "type": "java.lang.String",
+    "description": "A description for 'ucs.pip-list[1].additional-properties.KEY'"
+  },
+  {
+    "name": "ucstest.policy-file",
+    "type": "java.lang.String",
+    "description": "A description for 'ucstest.policy-file'"
+  }
+]}
