+++ UCSCore/src/test/java/it/cnr/iit/ucsrest/properties/UCSRestObligationManagerProperties.java
@@ -1,0 +1,7 @@
+package it.cnr.iit.ucsrest.properties;
+
+import it.cnr.iit.ucs.properties.components.ObligationManagerProperties;
+
+public class UCSRestObligationManagerProperties extends BaseProperties implements ObligationManagerProperties {
+
+}
