+++ UCSCore/src/test/java/it/cnr/iit/ucsrest/properties/UCSRestPdpProperties.java
@@ -1,0 +1,21 @@
+package it.cnr.iit.ucsrest.properties;
+
+import org.springframework.beans.factory.annotation.Value;
+
+import it.cnr.iit.ucs.properties.components.PdpProperties;
+
+public class UCSRestPdpProperties extends BaseProperties implements PdpProperties {
+
+    @Value( "${journal-dir}" )
+    private String journalDir;
+
+    @Override
+    public String getJournalDir() {
+        return journalDir;
+    }
+
+    public void setJournalDir( String journalDir ) {
+        this.journalDir = journalDir;
+    }
+
+}
