+++ UCSCore/src/test/resources/application.properties
