+++ UCSCore/src/test/resources/pips/camera.txt
@@ -1,0 +1,1 @@
+0
