+++ UCSCore/src/test/resources/pips/position.txt
@@ -1,0 +1,1 @@
+Alessandro	Pisa
