+++ UCSCore/src/test/resources/pips/role.txt
@@ -1,0 +1,1 @@
+Alessandro	IIT
