+++ UCSCore/src/test/resources/pips/telephone.txt
@@ -1,0 +1,1 @@
+0
