+++ UCSCore/src/test/resources/pips/temperature.txt
@@ -1,0 +1,1 @@
+27.0
