+++ UCSCore/src/test/resources/pips/time.txt
@@ -1,0 +1,1 @@
+7:00
