+++ UCSCore/src/test/resources/pips/virus.txt
@@ -1,0 +1,1 @@
+0
