+++ UCSCore/src/test/resources/xmls/request_1.xml
@@ -1,0 +1,43 @@
+<Request xmlns="urn:oasis:names:tc:xacml:3.0:core:schema:wd-17"
+	CombinedDecision="false" ReturnPolicyIdList="false">
+
+	<Attributes
+		Category="urn:oasis:names:tc:xacml:1.0:subject-category:access-subject">
+		<Attribute IncludeInResult="false"
+			AttributeId="urn:oasis:names:tc:xacml:1.0:subject:subject-id">
+			<AttributeValue
+				DataType="http://www.w3.org/2001/XMLSchema#string">Alessandro</AttributeValue>
+		</Attribute>
+	</Attributes>
+
+	<Attributes
+		Category="urn:oasis:names:tc:xacml:3.0:attribute-category:environment">
+		<Attribute
+			AttributeId="urn:oasis:names:tc:xacml:1.0:environment:environment-id"
+			IncludeInResult="false">
+			<AttributeValue
+				DataType="http://www.w3.org/2001/XMLSchema#string">environment</AttributeValue>
+		</Attribute>
+	</Attributes>
+
+	<Attributes
+		Category="urn:oasis:names:tc:xacml:3.0:attribute-category:action">
+		<Attribute
+			AttributeId="urn:oasis:names:tc:xacml:1.0:action:action-id"
+			IncludeInResult="false">
+			<AttributeValue
+				DataType="http://www.w3.org/2001/XMLSchema#string">publish</AttributeValue>
+		</Attribute>
+	</Attributes>
+
+	<Attributes
+		Category="urn:oasis:names:tc:xacml:3.0:attribute-category:resource">
+		<Attribute
+			AttributeId="urn:oasis:names:tc:xacml:1.0:resource:resource-id"
+			IncludeInResult="false">
+			<AttributeValue
+				DataType="http://www.w3.org/2001/XMLSchema#string">aaa</AttributeValue>
+		</Attribute>
+	</Attributes>
+
+</Request>
