+++ UCSRest/src/main/java/it/cnr/iit/ucsrest/rest/UCSRestStarter.java
@@ -28,7 +28,12 @@ import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.service.ApiInfo;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger.web.DocExpansion;
+import springfox.documentation.swagger.web.ModelRendering;
+import springfox.documentation.swagger.web.OperationsSorter;
+import springfox.documentation.swagger.web.TagsSorter;
 import springfox.documentation.swagger.web.UiConfiguration;
+import springfox.documentation.swagger.web.UiConfigurationBuilder;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 /**
@@ -56,7 +61,22 @@ public class UCSRestStarter extends SpringBootServletInitializer {
 
     @Bean
     public UiConfiguration uiConfig() {
-        return UiConfiguration.DEFAULT;
+        return UiConfigurationBuilder.builder()
+            .deepLinking( true )
+            .displayOperationId( false )
+            .defaultModelsExpandDepth( 1 )
+            .defaultModelExpandDepth( 1 )
+            .defaultModelRendering( ModelRendering.EXAMPLE )
+            .displayRequestDuration( false )
+            .docExpansion( DocExpansion.NONE )
+            .filter( false )
+            .maxDisplayedTags( null )
+            .operationsSorter( OperationsSorter.ALPHA )
+            .showExtensions( false )
+            .tagsSorter( TagsSorter.ALPHA )
+            .supportedSubmitMethods( UiConfiguration.Constants.DEFAULT_SUBMIT_METHODS )
+            .validatorUrl( null )
+            .build();
     }
 
     private ApiInfo metadata() {
