+++ UCSRest/src/main/resources/application.properties
@@ -3,10 +3,10 @@ server.port=9998
 ucs.core.uri=http://localhost:9998
 ucs.core.journal-dir=/tmp/ucf
 
-ucs.context-handler.name=it.cnr.iit.ucsrest.contexthandler.ContextHandler
+ucs.context-handler.name=it.cnr.iit.ucs.contexthandler.ContextHandler
 ucs.context-handler.uri=http://localhost:9998
 
-ucs.request-manager.name=it.cnr.iit.ucsrest.requestmanager.RequestManager
+ucs.request-manager.name=it.cnr.iit.ucs.requestmanager.RequestManager
 ucs.request-manager.api-remote-response=/retrieveRemoteResponse
 ucs.request-manager.active=true
 
@@ -21,7 +21,7 @@ ucs.policy-administration-point.path=
 
 ucs.obligation-manager.name=it.cnr.iit.ucs.obligationmanager.ObligationManager
 
-ucs.pep-list[0].name=it.cnr.iit.ucsrest.proxies.PEPRestProxy
+ucs.pep-list[0].name=it.cnr.iit.ucs.proxies.PEPRestProxy
 ucs.pep-list[0].id=0
 ucs.pep-list[0].uri=http://localhost:9999
 ucs.pep-list[0].revoke-type=SOFT
