+++ pom.xml
@@ -1,9 +1,9 @@
-<!-- Copyright 2018 IIT-CNR Licensed under the Apache License, Version 2.0
-	(the "License"); you may not use this file except in compliance with the
-	License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
-	Unless required by applicable law or agreed to in writing, software distributed
-	under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
-	OR CONDITIONS OF ANY KIND, either express or implied. See the License for
+<!-- Copyright 2018 IIT-CNR Licensed under the Apache License, Version 2.0 
+	(the "License"); you may not use this file except in compliance with the 
+	License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0 
+	Unless required by applicable law or agreed to in writing, software distributed 
+	under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES 
+	OR CONDITIONS OF ANY KIND, either express or implied. See the License for 
 	the specific language governing permissions and limitations under the License. -->
 <project xmlns="http://maven.apache.org/POM/4.0.0"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
@@ -25,6 +25,7 @@
 		<module>PolicyDecisionPoint</module>
 		<module>SessionManager</module>
 		<module>UCSRest</module>
+		<module>UCSCore</module>
 		<module>PEPRest</module>
 	</modules>
